Create scripts/painting.lua
